Rename utils.ts to parse.ts and order it sensibly
`utils` said nothing about what the module does. It is not a grab bag: every
one of its ten exports converts an HTML attribute value into a value the engine
expects, which makes it the module every attributeChangedCallback in the library
depends on. `parse.ts` says that.

getEntity looked like the exception that would block the rename, but it is not.
All twelve of its call sites - across button, script, scrollbar and scrollview
components - resolve an attribute string to an entity, exactly as the parsers
resolve one to a colour or a vector. Its name is kept, though: no default, no
attribute name, no warning, and null rather than a fallback, so the different
prefix communicates a genuinely different contract.

The order was historical accretion - neither alphabetical, nor grouped, nor
dependency-ordered. parseEnum and parseNumber were stranded after the Vec
family despite parseNumber being the most-called helper in the library; the two
shared internals sat in the middle, after two parsers that do not use them; and
parseTags was second because that is where I happened to add it. Now three
buckets, alphabetical within each: internals, then parsers, then the resolver.
Reading top-down you meet parseComponents and cloneDefault before their users.

Also adds a module header stating the contract the parsers share - null yields
the default, malformed yields the default plus exactly one warning, math
defaults are cloned, and the two helpers that cannot fail take no attribute
name. That was previously only inferable by reading nine JSDoc blocks.

Mechanical parts: every import site updated, and the test module renamed to
test/unit/parse.test.ts to keep mirroring its source. Both moves are staged as
renames so history follows.

One non-obvious follow-through: custom-elements-manifest.config.mjs excluded
src/utils.ts by path. Left alone it would have started feeding parse.ts to the
analyzer and changed the shipped manifest. Verified after the change that the
manifest still has 30 modules, still validates 27 elements, and contains no
module for parse.

No public API change - the module is internal and was never re-exported from
index.ts, so nothing in dist/index.d.ts moves.
