+++ custom-elements-manifest.config.mjs
@@ -10,7 +10,7 @@ export default {
     // Neither file declares an element, so excluding them keeps the manifest to the public
     // element surface. The base classes in async-element.ts and components/component.ts are
     // deliberately included - the elements that extend them inherit their attributes and events.
-    exclude: ['src/colors.ts', 'src/utils.ts'],
+    exclude: ['src/colors.ts', 'src/parse.ts'],
 
     // dist is wiped by `prebuild` and shipped via the `files` and `exports` fields, so the
     // generated files can never go stale and never need committing
